+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'morae'
+rootProject.name = 'nadab'
