--- src/main/java/com/devkor/ifive/morae/MoraeApplication.java
@@ -1,13 +1,0 @@
-package com.devkor.ifive.morae;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MoraeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MoraeApplication.class, args);
-	}
-
-}
