--- src/main/java/com/devkor/ifive/morae/domain/user/core/entity/SignupStatusType.java
@@ -1,7 +1,0 @@
-package com.devkor.ifive.morae.domain.user.core.entity;
-
-public enum SignupStatusType {
-    PROFILE_INCOMPLETE,  // 프로필 정보 미입력
-    COMPLETED,           // 가입 완료
-    WITHDRAWN            // 회원 탈퇴
-}
