--- src/main/java/com/devkor/ifive/morae/global/core/properties/CorsProperties.java
@@ -1,16 +1,0 @@
-package com.devkor.ifive.morae.global.core.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-@Getter
-@Setter
-@ConfigurationProperties(prefix = "cors")
-public class CorsProperties {
-    private List<String> allowedOrigins;
-}
