--- src/main/java/com/devkor/ifive/morae/global/exception/ErrorResponse.java
@@ -1,12 +1,0 @@
-package com.devkor.ifive.morae.global.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ErrorResponse {
-
-    private final int status;
-    private final String message;
-}
