--- src/main/java/com/devkor/ifive/morae/global/exception/JwtAuthException.java
@@ -1,10 +1,0 @@
-package com.devkor.ifive.morae.global.exception;
-
-/**
- * JWT (Access token) 검증 예외
- */
-public class JwtAuthException extends UnauthorizedException {
-    public JwtAuthException(String message) {
-        super(message);
-    }
-}
