--- src/main/java/com/devkor/ifive/morae/global/shared/Auditable.java
@@ -1,34 +1,0 @@
-package com.devkor.ifive.morae.global.shared;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.MappedSuperclass;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
-import lombok.Getter;
-
-import java.time.OffsetDateTime;
-
-@MappedSuperclass
-@Getter
-public abstract class Auditable {
-
-    @Column(name = "created_at", updatable = false)
-    protected OffsetDateTime createdAt;
-
-    @Column(name = "updated_at")
-    protected OffsetDateTime updatedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = OffsetDateTime.now();
-        updatedAt = createdAt;
-        postOnCreate();
-    }
-
-    protected void postOnCreate() {}
-
-    @PreUpdate
-    protected void onUpdate() {
-        updatedAt = OffsetDateTime.now();
-    }
-}
