+++ src/main/java/com/devkor/ifive/nadab/NadabApplication.java
@@ -1,0 +1,13 @@
+package com.devkor.ifive.nadab;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NadabApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NadabApplication.class, args);
+	}
+
+}
