+++ src/main/java/com/devkor/ifive/nadab/domain/auth/core/repository/RefreshTokenRepository.java
@@ -1,0 +1,22 @@
+package com.devkor.ifive.nadab.domain.auth.core.repository;
+
+import com.devkor.ifive.nadab.domain.auth.core.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Optional<RefreshToken> findByHashedToken(String hashedToken);
+
+    @Modifying
+    @Query("DELETE FROM RefreshToken rt WHERE rt.hashedToken = :hashedToken")
+    void deleteByHashedToken(@Param("hashedToken") String hashedToken);
+
+    @Modifying
+    @Query("DELETE FROM RefreshToken rt WHERE rt.user.id = :userId")
+    int deleteByUserId(@Param("userId") Long userId);
+}
