+++ src/main/java/com/devkor/ifive/nadab/global/exception/ErrorResponse.java
@@ -1,0 +1,12 @@
+package com.devkor.ifive.nadab.global.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorResponse {
+
+    private final int status;
+    private final String message;
+}
