+++ src/main/java/com/devkor/ifive/nadab/global/security/handler/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,39 @@
+package com.devkor.ifive.nadab.global.security.handler;
+
+import com.devkor.ifive.nadab.global.exception.ErrorResponse;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+/**
+ * JWT 인증 실패 시 401 Unauthorized 응답 처리
+ * - 인증되지 않은 사용자가 보호된 리소스 접근 시
+ * - 예: JWT 토큰이 없거나, 만료되었거나, 유효하지 않을 때
+ */
+@Component
+@RequiredArgsConstructor
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void commence(HttpServletRequest request,
+                         HttpServletResponse response,
+                         AuthenticationException authException) throws IOException {
+
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setCharacterEncoding("UTF-8");
+
+        ErrorResponse errorResponse = new ErrorResponse(HttpServletResponse.SC_UNAUTHORIZED, "JWT 인증에 실패했습니다.");
+
+        response.getWriter().write(objectMapper.writeValueAsString(errorResponse));
+    }
+}
