+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: morae
+    name: nadab
 
   profiles:
     active: local
