--- src/test/java/com/devkor/ifive/morae/MoraeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devkor.ifive.morae;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MoraeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
