+++ src/test/java/com/devkor/ifive/nadab/NadabApplicationTests.java
@@ -1,0 +1,13 @@
+package com.devkor.ifive.nadab;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NadabApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
